# docs/swagger.yaml
definitions:
  dto.DefaultHttpResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      ok:
        type: boolean
      statusCode:
        type: integer
    type: object
  schema.CmsConfigs:
    properties:
      backgroundColor:
        type: string
      companyCode:
        type: string
      companyLogoUrl:
        type: string
      createdAt:
        type: string
      display:
        type: string
      id:
        type: string
      primaryColor:
        type: string
      scope:
        type: string
      themeColor:
        type: string
    required:
    - companyCode
    type: object
host: companycms.nukeapp.dev
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Company Config Engine
  version: "1.0"
paths:
  /cms:
    get:
      consumes:
      - application/json
      description: get all cms config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.CmsConfigs'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
      security:
      - BearerAuth: []
      summary: get all cms config
      tags:
      - CMS Company
  /cms/{companyCode}:
    get:
      consumes:
      - application/json
      description: get one cms config
      parameters:
      - description: Company Code
        in: path
        name: companyCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.CmsConfigs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.DefaultHttpResponse'
      security:
      - BearerAuth: []
      summary: get one cms config
      tags:
      - CMS Company
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"